[avalanche] Fix inconsistent numeric types
Summary: Method arguments are aligned type-wise with their member type counterparts.

Test Plan: did rebuild and rerun unit tests

Reviewers: O1 Bitcoin ABC, #bitcoin_abc, deadalnix, jasonbcox

Reviewed By: O1 Bitcoin ABC, #bitcoin_abc, jasonbcox

Subscribers: teamcity, schancel

Differential Revision: https://reviews.bitcoinabc.org/D2174

+++ src/avalanche.cpp
@@ -247,7 +247,7 @@ bool AvalancheProcessor::registerVotes(
     return true;
 }
 
-bool AvalancheProcessor::addPeer(NodeId nodeid, uint32_t score) {
+bool AvalancheProcessor::addPeer(NodeId nodeid, int64_t score) {
     return peerSet.getWriteView()
         ->insert({nodeid, score, std::chrono::steady_clock::now()})
         .second;
